MAJOR CHANGE
Changed gameplay to instead of checking for the correct pose after 2 seconds, checking constantly for the correct pose.

Scoring is now time-based. The game keeps track and displays how long it took the player to make the correct pose. A final total of all times is displayed at the end of the game.